YES! it gets through the game, recieving moves and sending the player updated info going from preflop to flop, and so on, looping the game to start a new flop! only really tested with calling, haven't tried raising or folding, but from what I understand the graphics aren't updated to handle that.
Still need to tell the client when it's it's turn, and notify the user,
but as far as I can tell the game should work once we get that and
scoring. also need to make sure to update the actual render when the
table changes. I'm probably speaking too soon about things working but oh well

also I took out the timer, it just was annoying to code around